backend/requirements.txt
django
psycopg2-binary
django-environ
celery
django-celery-results
djangorestframework
django-cors-headers
pika
django-environ
